refactor: Create and test Book component
BookSpecification is a class that describes
a book.

BookList is a class that contains an array of
BookSpecification objects. The constructor
defines BookSpecfications that define the seven
given books.

Book is a class that contains a method to
render a description (title, author, image of
cover, menu for moving the book from one shelf
to another) on the screen.