feat: Add a means to filter a list of books
Classify books. Some books are currently being read,
some books have already been read, and some books will be
read in the future.

Create a means to select books in one of the three
categories from the list of all books.

Create a class named When to simulate the Enum types
of Java. How do other JavaScript programmers do this?