feat: Display selected books on search page
Create a BookShelf on the search page to hold the
results of searchs.

Write the code that reads a search string, uses that
string to query the database, and stores the results
in an array.

Check for names of authors and images of the books'
covers---these may be missing in some records.

In cases where a book has no author, place the words
"No author" in the author field.

In cases where a record does not contain the URL for
an image of the book's cover, place a URL for an
image of the Boston skyline in the record---this will
be the default image.