feat: Make selected books visible on main page
I did a lot of work and work of several different kinds
since the last commit.

I added a mean of making books that I select on the search
page visible on the main page. A selection identities a book
and a shelf.

Each return to the main page from the search page duplicated
the list of books to be displayed. I solved this by searching
for duplicates before adding a book to the list. I think that
a better solution would empty the list before reading from the
database and then rebuild the list. My attempt to solve the
problem that way did not work because my code was not executed
in the order that I expected. I need to improve my understanding
of setState() and asynchronous execution.

I have added PropTypes, converted classes that extend Component
to functions where that was possible, and added JSDoc comments.

I removed trailing spaces.