fix: Clear search results when given empty or unrecognized term
The program now deletes the results of the previous search
when a reader deletes the text in the search field.

The program now shows no results when a reader enters
text in the search field that the application does not
recognize. In the previous region, it retained the results
of the previous search.