fix: Show shelf of already selected books on search page
The menu with which readers select a shelf on which to place
a book now makes the shelf (if any) on which the book currently sits
the default choice.

To do this, I call BooksAPI.getAll() in the search page. I found that
BooksAPI.search() does not return the records that I placed in the
database with BooksAPI.update() on the main page. The records returned
by BooksAPI.search() do not contain the shelf information that
I included in calls to BooksAPI.update() and that I see in records
returned by BooksAPI.getAll(). I am confused. This is not what I
expected.